Move connection handler into its own file
